# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/currency/rate": {
            "get": {
                "description": "填入 \"來源幣別\"、\"目標幣別\"、\"金額數字\"，計算出當前匯率結果。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Currency"
                ],
                "summary": "計算匯率",
                "parameters": [
                    {
                        "enum": [
                            "TWD",
                            "JPY",
                            "USD"
                        ],
                        "type": "string",
                        "description": "來源幣別",
                        "name": "from",
                        "in": "query",
                        "required": true
                    },
                    {
                        "enum": [
                            "TWD",
                            "JPY",
                            "USD"
                        ],
                        "type": "string",
                        "description": "目標幣別",
                        "name": "to",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "number",
                        "default": 0,
                        "description": "金額數字",
                        "name": "amount",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"res\": \"3.67\"}",
                        "schema": {
                            "$ref": "#/definitions/currency_rate.RateRes"
                        }
                    },
                    "400": {
                        "description": "{\"message\": \"invalid input\", \"error\": \"req.from can not be null\"}",
                        "schema": {
                            "$ref": "#/definitions/currency_rate.RateRes"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "currency_rate.RateRes": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "res": {
                    "type": "string"
                }
            }
        }
    }
}